requirements.txt
pyyaml==5.3
nltk==3.4.5